Refactored functions into seperate files#1
