+++ server/index.js
@@ -1,8 +1,17 @@
 require('dotenv/config')
 const { createApp } = require('./create-app')
+const { MongoClient } = require('mongodb')
+const uploadsGateway = require('./uploads-gateway')
 
-const app = createApp()
+MongoClient.connect(process.env.MONGODB_URI, async (err, db) => {
 
-app.listen(process.env.PORT, () => {
-  console.log('Listening on ' + process.env.PORT)
+  if (err) console.log(err)
+
+  const collection = db.collection('uploads')
+  const app = createApp(uploadsGateway(collection))
+
+  app.listen(process.env.PORT, () => {
+    console.log('Listening on ' + process.env.PORT)
+  })
+
 })
