+++ server/uploads-gateway.js
@@ -1,0 +1,7 @@
+module.exports = function uploadsGateway(collection) {
+  return {
+    async create(uploadData) {
+      collection.insertOne(uploadData)
+    }
+  }
+}
